requirements.txt
flask
Pillow
brother_ql
